services/knowledge-graph/requirements.txt
networkx
strawberry-graphql
uvicorn
fastapi
pyyaml
neo4j
python-dotenv
